# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Actual Bridge API
  version: 1.0.0
  description: |
    Stable contract between the Android app and the Actual bridge.

    Conventions
    - All money values are integers in minor units (cents). Ledger amounts are signed:
      outflows negative, inflows positive. Budget amounts (`budgeted`, `totalBudgeted`)
      are the amount assigned, positive when money is assigned.
    - Months are `YYYY-MM`, dates are `YYYY-MM-DD`.
    - Within /v1 changes are additive only. Clients must ignore unknown fields and
      treat unknown enum values as `unknown`.
    - Non-idempotent writes require an `Idempotency-Key` header. Creates use a
      client-generated UUID `id` instead.
    - Errors are `application/problem+json` with a stable `code`.
servers:
  - url: https://budget-api.example.com
security:
  - bearer: []
...
paths:
...
  /v1/budgets/{budgetId}/schedules:
    get:
      tags: [budgets]
      operationId: listSchedules
      description: Scheduled and recurring transactions (Monarch "Recurring").
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: upcoming, in: query, description: How many upcoming dates to include, schema: { type: integer, minimum: 1, maximum: 24, default: 3 } }
      responses:
        '200':
          description: Schedules
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Schedule' } } } }
        '423': { $ref: '#/components/responses/Problem' }
    post:
      tags: [budgets]
      operationId: createSchedule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/schedules/{id}:
    patch:
      tags: [budgets]
      operationId: updateSchedule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [budgets]
      operationId: deleteSchedule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/schedules/{id}/skip:
    post:
      tags: [budgets]
      operationId: skipSchedule
      description: Skip the next occurrence. Capability `schedules.skip`.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '404': { $ref: '#/components/responses/Problem' }
        '501': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/schedules/{id}/post:
    post:
      tags: [budgets]
      operationId: postSchedule
      description: Add the next occurrence as a transaction now. Capability `schedules.post`.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Schedule' }
        '404': { $ref: '#/components/responses/Problem' }
        '501': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/rules:
    get:
      tags: [budgets]
      operationId: listRules
      description: All rules, including the hidden ones behind schedules (scheduleId set).
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '200':
          description: Rules
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Rule' } } } }
        '423': { $ref: '#/components/responses/Problem' }
    post:
      tags: [budgets]
      operationId: createRule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rule' }
        '400': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/rules/{id}:
    put:
      tags: [budgets]
      operationId: updateRule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RuleInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rule' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [budgets]
      operationId: deleteRule
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/payees/stats:
    get:
      tags: [budgets]
      operationId: listPayeeStats
      description: Payees with usage counts (Monarch "Merchants").
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '200':
          description: Payees
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/PayeeStat' } } } }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/payees/{id}:
    patch:
      tags: [budgets]
      operationId: renamePayee
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [name], properties: { name: { type: string, minLength: 1, maxLength: 200 } } }
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payee' }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [budgets]
      operationId: deletePayee
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Deleted }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/payees/{id}/merge:
    post:
      tags: [budgets]
      operationId: mergePayees
      description: Moves transactions and rules from mergeIds onto this payee, then removes them.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [mergeIds], properties: { mergeIds: { type: array, minItems: 1, items: { type: string } } } }
      responses:
        '204': { description: Merged }
        '400': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/tags:
    get:
      tags: [budgets]
      operationId: listTags
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      responses:
        '200':
          description: Tags
          content:
            application/json:
              schema: { type: object, required: [items], properties: { items: { type: array, items: { $ref: '#/components/schemas/Tag' } } } }
        '423': { $ref: '#/components/responses/Problem' }
    post:
      tags: [budgets]
      operationId: createTag
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '409': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/tags/{id}:
    patch:
      tags: [budgets]
      operationId: updateTag
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagInput' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tag' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [budgets]
      operationId: deleteTag
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/categories/{id}/note:
    get:
      tags: [budgets]
      operationId: getCategoryNote
      description: 'Category notes hold #template goal lines.'
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Note
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryNote' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
    put:
      tags: [budgets]
      operationId: setCategoryNote
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [note], properties: { note: { type: [string, 'null'] } } }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CategoryNote' }
        '404': { $ref: '#/components/responses/Problem' }
        '423': { $ref: '#/components/responses/Problem' }
...
  /v1/budgets/{budgetId}/rules/preview:
    post:
      tags: [budgets]
      operationId: previewRule
      summary: "Which transactions a rule (saved or not) matches, and what it would change on the first `limit` of them. Nothing is written. Matching is Actual's own; formula results are worked out like Actual does (numbers out are currency units); BALANCE_OF, QUERY and BUDGET_QUERY only when the rule runs."
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [conditions, actions]
              properties:
                conditionsOp: { type: string, enum: [and, or] }
                conditions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
                actions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
                stage: { type: [string, 'null'] }
                limit: { type: integer, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: Matches and changes
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RulePreview' }
  /v1/budgets/{budgetId}/rules/{id}/run:
    post:
      tags: [budgets]
      operationId: runRule
      summary: Members. Runs a saved rule's actions on the existing transactions it matches (or just `transactionIds`, where they match).
      parameters:
        - $ref: '#/components/parameters/BudgetId'
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { transactionIds: { type: array, items: { type: string } } } }
      responses:
        '200':
          description: How many were updated
          content:
            application/json:
              schema: { type: object, required: [updated], properties: { updated: { type: integer } } }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/rules/rerun:
    post:
      tags: [budgets]
      operationId: rerunRules
      summary: Members. Runs every rule again on these transactions, as if they had just been imported.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [transactionIds], properties: { transactionIds: { type: array, minItems: 1, maxItems: 500, items: { type: string } } } }
      responses:
        '200':
          description: How many changed
          content:
            application/json:
              schema: { type: object, required: [checked, changed], properties: { checked: { type: integer }, changed: { type: integer } } }
        '403': { $ref: '#/components/responses/Problem' }
  /v1/budgets/{budgetId}/rules/run-all:
    post:
      tags: [budgets]
      operationId: runAllRules
      summary: Members. Runs every rule again on every transaction since `since` (all of them without it), as if just imported. Reconciled transactions are left alone.
      parameters:
        - $ref: '#/components/parameters/BudgetId'
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { since: { type: [string, 'null'], pattern: '^\d{4}-\d{2}-\d{2}$' } } }
      responses:
        '200':
          description: How many were checked and changed
          content:
            application/json:
              schema: { type: object, required: [checked, changed], properties: { checked: { type: integer }, changed: { type: integer } } }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    BudgetId:
      name: budgetId
      in: path
      required: true
      description: Actual sync id (the budget's `groupId`), stable across devices
      schema: { type: string }
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Money:
      type: integer
      format: int64
      description: Minor units (cents)
...
    Date:
      type: string
      pattern: '^\d{4}-\d{2}-\d{2}$'
...
    Payee:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        transferAccountId:
          type: [string, 'null']
          description: Set when this payee represents a transfer to that account
...
    RulePreview:
      type: object
      required: [matchCount, items, errors]
      properties:
        matchCount: { type: integer, description: Every existing transaction the conditions match }
        errors: { type: array, items: { type: string }, description: Conditions Actual couldn't turn into a search }
        items:
          type: array
          items:
            type: object
            required: [transaction, changes]
            properties:
              transaction: { $ref: '#/components/schemas/Transaction' }
              changes:
                type: array
                items:
                  type: object
                  required: [field, value]
                  properties:
                    field: { type: string }
                    value: { description: "The new value (ids shown as names; amounts in minor units)" }
                    note: { type: string }
                    error: { type: string }
...
    Schedule:
      type: object
      required: [id, name, nextDate, completed, postsTransaction, payeeId, accountId, amount, amountMax, amountOp, recurrence, date, upcoming]
      properties:
        id: { type: string }
        name: { type: [string, 'null'] }
        nextDate: { type: [string, 'null'] }
        completed: { type: boolean }
        postsTransaction: { type: boolean, description: Actual adds the transaction automatically when due }
        payeeId: { type: [string, 'null'] }
        accountId: { type: [string, 'null'] }
        amount: { $ref: '#/components/schemas/Money' }
        amountMax: { type: [integer, 'null'], description: Upper bound when amountOp is isbetween }
        amountOp: { type: string, description: "is, isapprox or isbetween" }
        recurrence:
          oneOf:
            - { $ref: '#/components/schemas/Recurrence' }
            - { type: 'null' }
        date: { type: [string, 'null'], description: One-time schedules only }
        upcoming:
          type: array
          items: { type: string }
    ScheduleInput:
      type: object
      minProperties: 1
      properties:
        name: { type: [string, 'null'] }
        payeeId: { type: [string, 'null'] }
        payeeName: { type: string }
        accountId: { type: [string, 'null'] }
        amount: { $ref: '#/components/schemas/Money' }
        amountMax: { $ref: '#/components/schemas/Money' }
        amountOp: { type: string, enum: [is, isapprox, isbetween] }
        recurrence: { $ref: '#/components/schemas/Recurrence' }
        date: { $ref: '#/components/schemas/Date' }
        postsTransaction: { type: boolean }
    RuleItem:
      type: object
      description: "A condition or action in Actual's rule format; value's type depends on field and op"
      required: [op]
      properties:
        field: { type: [string, 'null'] }
        op: { type: string }
        value: {}
        type: { type: [string, 'null'] }
        options: { type: [object, 'null'] }
    Rule:
      type: object
      required: [id, stage, conditionsOp, conditions, actions, scheduleId]
      properties:
        id: { type: string }
        stage: { type: [string, 'null'], description: "pre, post or null (default)" }
        conditionsOp: { type: string, enum: [and, or] }
        conditions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
        actions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
        scheduleId: { type: [string, 'null'], description: Set for the hidden rule behind a schedule }
    RuleInput:
      type: object
      required: [conditions, actions]
      properties:
        stage: { type: [string, 'null'], enum: [pre, post, null] }
        conditionsOp: { type: string, enum: [and, or] }
        conditions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
        actions: { type: array, items: { $ref: '#/components/schemas/RuleItem' } }
    PayeeStat:
      type: object
      required: [id, name, transferAccountId, transactionCount, ruleCount]
      properties:
        id: { type: string }
        name: { type: string }
        transferAccountId: { type: [string, 'null'] }
        transactionCount: { type: integer }
        ruleCount: { type: integer }
    Tag:
      type: object
      required: [id, tag, color, description]
      properties:
        id: { type: string }
        tag: { type: string, description: 'Without the leading #' }
        color: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
    TagInput:
      type: object
      minProperties: 1
      properties:
        tag: { type: string }
        color: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
    CategoryNote:
      type: object
      required: [categoryId, note]
      properties:
        categoryId: { type: string }
        note: { type: [string, 'null'] }
...